Feat : Modularizing the app.py file
- place.py : function that save place and listing the place
- auth.py : sign in, sign out, sign up function and control the user session
- map.py : function that load the place list and mark the marker
