+++ templates/login.html
@@ -39,6 +39,7 @@
 
     <!-- Import JS -->
     <script src="{{ url_for('static', filename='js/login.js') }}"></script>
+    <script src="{{ url_for('static', filename='js/register.js') }}"></script>
 
 
     <style>
